adding route and date picker
