+++ src/pages/formulaires.js
@@ -1,11 +1,11 @@
 import React from "react"
 import { Router } from "@reach/router"
 import Layout from "../components/layout"
-import PageForm from "../components/journal_de_bord_psh"
+import JdbPshForm from "../components/forms/jdb_psh"
 const App = () => (
-  <Layout>
+  <Layout role="main">
     <Router>
-      <PageForm path="/form/jdb" />
+      <JdbPshForm path="/formulaires/journal-de-bord/p/:andi_id" />
     </Router>
   </Layout>
 )
